Organized and refactored code
