+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/features/narrator_menu/NarratorMenu.java
@@ -4,7 +4,7 @@ import com.github.khanshoaib3.minecraft_access.MainClass;
 import com.github.khanshoaib3.minecraft_access.features.BiomeIndicator;
 import com.github.khanshoaib3.minecraft_access.features.ReadCrosshair;
 import com.github.khanshoaib3.minecraft_access.utils.ClientPlayerEntityUtils;
-import com.github.khanshoaib3.minecraft_access.utils.PlayerPosition;
+import com.github.khanshoaib3.minecraft_access.utils.PositionUtils;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.MinecraftClient;
@@ -16,8 +16,6 @@ import net.minecraft.text.MutableText;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.hit.HitResult;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Direction;
-import net.minecraft.util.math.Vec3d;
 import net.minecraft.world.biome.Biome;
 
 // TODO add doc and update readme
@@ -75,7 +73,7 @@ public class NarratorMenu {
                         Block block = blockState.getBlock();
                         MutableText mutableText = block.getName();
 
-                        String text = mutableText.getString() + ", " + getPositionDifference(blockPos);
+                        String text = mutableText.getString() + ", " + PositionUtils.getPositionDifference(blockPos);
                         MainClass.speakWithNarrator(text, true);
                     } catch (Exception e) {
                         e.printStackTrace();
@@ -109,7 +107,7 @@ public class NarratorMenu {
                     try {
                         BlockHitResult blockHitResult = (BlockHitResult) hit;
                         BlockPos blockPos = blockHitResult.getBlockPos();
-                        MainClass.speakWithNarrator(getPosition(blockPos), true);
+                        MainClass.speakWithNarrator(PositionUtils.getPosition(blockPos), true);
                     } catch (Exception e) {
                         e.printStackTrace();
                     }
@@ -135,7 +133,7 @@ public class NarratorMenu {
                 name = name.replace("block.minecraft.", ""); // Remove `block.minecraft.` for unsupported languages
 
             if (onlyPosition) {
-                MainClass.speakWithNarrator(getPosition(blockPos), true);
+                MainClass.speakWithNarrator(PositionUtils.getPosition(blockPos), true);
                 return true;
             }
 
@@ -143,7 +141,7 @@ public class NarratorMenu {
             String levelString = "";
             if (level < 8) levelString = I18n.translate("minecraft_access.read_crosshair.fluid_level", level);
 
-            String toSpeak = name + levelString + ", " + getPositionDifference(blockPos);
+            String toSpeak = name + levelString + ", " + PositionUtils.getPositionDifference(blockPos);
 
             MainClass.speakWithNarrator(toSpeak, true);
             return true;
@@ -190,70 +188,5 @@ public class NarratorMenu {
         } catch (Exception e) {
             e.printStackTrace();
         }
-    }
-
-    public static String getPositionDifference(BlockPos blockPos) {
-        if (minecraftClient.player == null) return "up";
-
-        Direction dir = minecraftClient.player.getHorizontalFacing();
-
-//        Vec3d diff = minecraftClient.player.getEyePos().subtract(Vec3d.ofCenter(blockPos)); // post 1.18
-        Vec3d diff = new Vec3d(minecraftClient.player.getX(), minecraftClient.player.getEyeY(), minecraftClient.player.getZ()).subtract(Vec3d.ofCenter(blockPos)); // pre 1.18
-        BlockPos diffBlockPos = new BlockPos(Math.round(diff.x), Math.round(diff.y), Math.round(diff.z));
-
-        String diffXBlockPos = "";
-        String diffYBlockPos = "";
-        String diffZBlockPos = "";
-
-        if (diffBlockPos.getX() != 0) {
-            if (dir == Direction.NORTH) {
-                diffXBlockPos = getDifferenceString(diffBlockPos.getX(), "right", "left");
-            } else if (dir == Direction.SOUTH) {
-                diffXBlockPos = getDifferenceString(diffBlockPos.getX(), "left", "right");
-            } else if (dir == Direction.EAST) {
-                diffXBlockPos = getDifferenceString(diffBlockPos.getX(), "away", "behind");
-            } else if (dir == Direction.WEST) {
-                diffXBlockPos = getDifferenceString(diffBlockPos.getX(), "behind", "away");
-            }
-        }
-
-        if (diffBlockPos.getY() != 0) {
-            diffYBlockPos = getDifferenceString(diffBlockPos.getY(), "up", "down");
-        }
-
-        if (diffBlockPos.getZ() != 0) {
-            if (dir == Direction.SOUTH) {
-                diffZBlockPos = getDifferenceString(diffBlockPos.getZ(), "away", "behind");
-            } else if (dir == Direction.NORTH) {
-                diffZBlockPos = getDifferenceString(diffBlockPos.getZ(), "behind", "away");
-            } else if (dir == Direction.EAST) {
-                diffZBlockPos = getDifferenceString(diffBlockPos.getZ(), "right", "left");
-            } else if (dir == Direction.WEST) {
-                diffZBlockPos = getDifferenceString(diffBlockPos.getZ(), "left", "right");
-            }
-        }
-
-        String text;
-        if (dir == Direction.NORTH || dir == Direction.SOUTH)
-            text = String.format("%s  %s  %s", diffZBlockPos, diffYBlockPos, diffXBlockPos);
-        else
-            text = String.format("%s  %s  %s", diffXBlockPos, diffYBlockPos, diffZBlockPos);
-        return text;
-    }
-
-    private static String getPosition(BlockPos blockPos) {
-        try {
-            String posX = PlayerPosition.getNarratableNumber(blockPos.getX());
-            String posY = PlayerPosition.getNarratableNumber(blockPos.getY());
-            String posZ = PlayerPosition.getNarratableNumber(blockPos.getZ());
-            return String.format("%s x %s y %s z", posX, posY, posZ);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return "";
-    }
-
-    private static String getDifferenceString(int blocks, String key1, String key2) {
-        return I18n.translate("minecraft_access.util.position_difference_" + (blocks < 0 ? key1 : key2), Math.abs(blocks));
     }
 }
